ai_arena/slack_msg_consumer/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
bcrypt==3.2.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator==5.0.7
google-api-core[grpc]==1.26.3
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-logging==2.3.1
google-cloud-pubsub==2.4.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos[grpc]==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.1
idna==2.10
libcst==0.3.18
mypy-extensions==0.4.3
numpy==1.20.2
packaging==21.3; python_version >= '3.6'
pandas==1.2.4
proto-plus==1.18.1
protobuf==3.15.8
psycopg2==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pyyaml==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
validators==0.18.2
